# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RideCompass API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/region/road-surface-tiles/{z}/{x}/{y}.pbf": {
      "get": {
        "operationId": "region_road_surface_tile_api_region_road_surface_tiles__z___x___y__pbf_get",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 15,
              "minimum": 12,
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/region/point-tiles/{layer}/{z}/{x}/{y}.pbf": {
      "get": {
        "operationId": "region_point_tile_api_region_point_tiles__layer___z___x___y__pbf_get",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 15,
              "minimum": 12,
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/region/landcover-tiles/{z}/{x}/{y}.png": {
      "get": {
        "operationId": "region_landcover_tile_api_region_landcover_tiles__z___x___y__png_get",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 14,
              "minimum": 6,
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/region/dynamic-way-values/{axis_id}/{z}/{x}/{y}": {
      "get": {
        "operationId": "region_dedicated_way_values_api_region_dynamic_way_values__axis_id___z___x___y__get",
        "parameters": [
          {
            "name": "axis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 15,
              "minimum": 12,
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Y"
            }
          },
          {
            "name": "bearing_deg",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMaximum": 360
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bearing Deg"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "name": "speed_kmh",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 60.0,
                  "minimum": 5.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Speed Kmh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response Region Dedicated Way Values Api Region Dynamic Way Values  Axis Id   Z   X   Y  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/region/axis-inspector": {
      "post": {
        "operationId": "region_axis_inspector_api_region_axis_inspector_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisInspectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AxisInspectorResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Region Axis Inspector Api Region Axis Inspector Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jma-tile/{path}": {
      "get": {
        "operationId": "jma_tile_proxy_api_jma_tile__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AxisInspectorAxis": {
        "properties": {
          "axis_id": {
            "type": "string",
            "title": "Axis Id"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "axis_id",
          "difficulty",
          "contribution"
        ],
        "title": "AxisInspectorAxis"
      },
      "AxisInspectorRequest": {
        "properties": {
          "osm_way_id": {
            "type": "integer",
            "title": "Osm Way Id"
          },
          "feature_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Key"
          },
          "z": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 12.0,
            "title": "Z"
          },
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "bearing_deg": {
            "type": "number",
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Bearing Deg"
          },
          "at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 60.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "route_preference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutePreferenceWeights"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "osm_way_id",
          "z",
          "x",
          "y",
          "bearing_deg"
        ],
        "title": "AxisInspectorRequest"
      },
      "AxisInspectorResult": {
        "properties": {
          "highway": {
            "type": "string",
            "title": "Highway"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          },
          "axes": {
            "items": {
              "$ref": "#/components/schemas/AxisInspectorAxis"
            },
            "type": "array",
            "title": "Axes"
          },
          "composite_difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InspectorComposite"
              },
              {
                "type": "null"
              }
            ]
          },
          "landcover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LandcoverPercentages"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "highway",
          "tags",
          "axes",
          "composite_difficulty",
          "landcover"
        ],
        "title": "AxisInspectorResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InspectorComposite": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value"
          },
          "covered_weight_fraction": {
            "type": "number",
            "title": "Covered Weight Fraction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value",
          "covered_weight_fraction"
        ],
        "title": "InspectorComposite"
      },
...
      "LandcoverPercentages": {
        "properties": {
          "water_percent": {
            "type": "number",
            "title": "Water Percent"
          },
          "trees_percent": {
            "type": "number",
            "title": "Trees Percent"
          },
          "flooded_veg_percent": {
            "type": "number",
            "title": "Flooded Veg Percent"
          },
          "crops_percent": {
            "type": "number",
            "title": "Crops Percent"
          },
          "built_percent": {
            "type": "number",
            "title": "Built Percent"
          },
          "bare_percent": {
            "type": "number",
            "title": "Bare Percent"
          },
          "snow_ice_percent": {
            "type": "number",
            "title": "Snow Ice Percent"
          },
          "rangeland_percent": {
            "type": "number",
            "title": "Rangeland Percent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "water_percent",
          "trees_percent",
          "flooded_veg_percent",
          "crops_percent",
          "built_percent",
          "bare_percent",
          "snow_ice_percent",
          "rangeland_percent"
        ],
        "title": "LandcoverPercentages"
      },
...
      "RoutePreferenceWeights": {
        "additionalProperties": {
          "type": "number"
        },
        "type": "object",
        "title": "RoutePreferenceWeights"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}